+++ packages/errors/src/types.ts
@@ -1,6 +1,6 @@
 // DEV REMINDER:
 // - update the README.md when this is changed
-export type ZeroDepErrorTax = 'type' | 'range' | 'reference' | 'syntax' | 'uri' | 'unknown';
+export type ZeroDepErrorCategory = 'type' | 'range' | 'reference' | 'syntax' | 'uri' | 'unknown';
 
 // DEV REMINDER:
 // - update the README.md when this is changed