+++ packages/guard/guard.errors/src/lib/ZeroDepErrorGuard.ts
@@ -1,13 +1,13 @@
-import { ZeroDepError, ZeroDepErrorSource, ZeroDepErrorTax } from '@zerodep/errors';
+import { ZeroDepError, ZeroDepErrorSource, ZeroDepErrorCategory } from '@zerodep/errors';
 
 export class ZeroDepErrorGuard extends ZeroDepError {
   constructor(
     message = 'Value is invalid',
-    tax: ZeroDepErrorTax = 'unknown',
+    category: ZeroDepErrorCategory = 'unknown',
     source: ZeroDepErrorSource = 'guard',
     value: any = undefined
   ) {
-    super(message, tax, source, value);
+    super(message, category, source, value);
 
     // ensure the name matches the class
     this.name = this.constructor.name;