+++ packages/app/CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Performance improvements to `is.*` algorithms
 - Fixed "How to Use" examples for arrays
+- ZeroDepError `tax` value renamed to `category` for clarity
 
 ## [0.1.0] - 2022-04-28