+++ packages/errors/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 
+- Internal `tax` value renamed to `category` for clarity
 - Fixed incorrect package description
 
 ## [0.1.3] - 2022-04-25