+++ packages/guard/guard.object/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 
+- Internal `tax` value renamed to `category` for clarity
 - Copy changes in the README
 
 ## [0.1.2] - 2022-04-28